Add a version requirement to inch_ex
Gets rid of a warning. Have the 0.4.0 version from the mix.lock,
although I wonder if this should be checked in? (At least for
libraries et.al. in Ruby the rule was to not check in .lock
which sort of gets you to be more vigilant with your specified
version requirements (which is what users will get when installing).
Alas, here it doesn't matter as much as there are only docs
requirements no runtime deps)