fix milli-/micro-second consistency
Prior to this commit the documentation indicated that the resolution of
timer triggers was in microseconds; however, the code, in all the
examples and with the `to_ms` function indicated that the timer
resolution was, in fact, milliseconds. This becomes clearer upon
evaluating the timer doctest example, where it uses a window of 1 hour.
1 hour is 3_600_000 milliseconds, and consistent with the orders of
magnitude for the data elements (3_200_000, 4_000_000); however, 1 hour
in microseconds would have been 3_600_000_000, in which case *all* of
example data would have fallen into the first window. This commit
simply updates the documentation to reflect the de facto millisecond
resolution as indicated in multiple aspects of the code.