Namespace ConsumerSupervisor as GenStage.ConsumerSupervisor
This PR is presented for comment - not namespacing this module caused me
some confusion when I upgraded an app from using `DynamicSupervisor`, as
I just assumed that it would be namespaced and was surprised when it
wasn't, perhaps because `DynamicSupervisor` was buried under the
`Experimental` namespace prior to this latest version. The more I've
thought about it the more I think it makes sense, but I'm interested to
see what others think.