Fix type for subscription_options
It seems like the name suggested it was a keyword list, but instead it
defined a single option. This confusion carried on into the spec for
`handle_subscribe` where the typespec defined the `opts` argument as a
tuple instead of a keyword list.

I opted for redefining the existing reference to this type since it
might be used outside as to not introduce any breaking changes. If this
isn't required, I'm open to renaming the existing type to
subscription_option and changing the existing spec to be a list of
tuples instead of a single tuple.