Log a warning on a bad subscription
Currently when you try to subscribe to a process that is not alive, the
error that is printed out does not tell you what process is not alive.
Rectify this by adding an additional log statement to cover this case.
This will help a developer figure out why their GenStage is not starting
and is especially helpful in the case that a Consumer is subscribing to
multiple different producers.