Do not duplicate demand in BroadcastDispatcher
Since 5d8193ee, the BroadcastDispatcher can send demand upstream even if
its consumers did not request that many events. Because `waiting` is set
to 0 when a new subscriber subscribes, the demand is repeated when the
new subscriber asks for events.

I am fixing this by introducing a new counter in state - the number of
requested but not yet delivered events. The dispatcher now does not send
upstream demand if the demand can be satisfied by the requested events
(and sends a lower demand if it can be satisfied partially).

Fixes #318.